moved api functionalities to token class, removed api file
